Refactor: use class for TokenRequestTwoFactor
